AAOS Coding Assignment
Use Case:

Use Case:

A car rental company wants notifications if the renter of a car drives at a speed above a given limit. The limit shall be set by the fleet company and it shall be possible for them to set the maximum permitted speed before  a rental period starts. It shall also be possible for different customers to have different maximum speeds set.
If the set speed is exceeded, the rental company shall be notified and the user given a warning alert.

The first communication channel will happen through Firebase. It is possible that we will need an alternative communication channel to AWS as well though.